Spanish translation (#984)
* Spanish translation

Some key words were not fully translated because they are commonly used and recognizable by people who have at least minimal exposure to terms related to artificial intelligence and stable diffusion, like 'steps', CFG, etc.

The term ‘Prompt’ remains unchanged because it is widely used in various contexts, and practically everyone is accustomed to hearing it. Additionally, there is no official translation for it. However, this may be subject to change if necessary."